+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/server/PekkoRoutesFactory.scala
@@ -74,7 +74,8 @@ object PekkoRoutesFactory {
         processingTypeServicesProvider.mapValues(_.scenarioValidator),
         processingTypeServicesProvider.mapValues(_.scenarioResolver),
         actionService,
-        processingTypeServicesProvider.mapValues(_.additionalComponentConfigs)
+        processingTypeServicesProvider.mapValues(_.additionalComponentConfigs),
+        limitsService
       )
       new ManagementResources(
         processAuthorizer,