+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/validation/NodeValidator.scala
@@ -4,8 +4,7 @@ import cats.effect.SyncIO
 import cats.effect.kernel.Resource
 import pl.touk.nussknacker.engine.{ModelData, ScenarioCompilationDependencies}
 import pl.touk.nussknacker.engine.api.{JobData, ProcessVersion}
-import pl.touk.nussknacker.engine.api.context.{ProcessCompilationError, ValidationContext}
-import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.MissingParameters
+import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.definition.EngineScenarioCompilationDependencies
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
 import pl.touk.nussknacker.engine.compile.FragmentResolver
@@ -67,16 +66,7 @@ class NodeValidator(
               )
             case ValidationPerformed(errors, parameters, expressionType) =>
               val uiParams = parameters.map(_.map(DefinitionsService.createUIParameter))
-
-              // We don't return MissingParameter error when we are returning those missing parameters to be added - since
-              // it's not really exception ATM
-              def shouldIgnoreError(pce: ProcessCompilationError): Boolean = pce match {
-                case MissingParameters(params, _) =>
-                  params.forall(missing => uiParams.exists(_.exists(_.name == missing.value)))
-                case _ => false
-              }
-
-              val uiErrors = errors.filterNot(shouldIgnoreError).map(PrettyValidationErrors.formatErrorMessage)
+              val uiErrors = errors.map(PrettyValidationErrors.formatErrorMessage)
               NodeValidationResult(
                 parameters = uiParams,
                 expressionType = expressionType,