+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/ProcessesExportImportResourcesSpec.scala
@@ -47,7 +47,7 @@ class ProcessesExportImportResourcesSpec
     futureFetchingScenarioRepository,
     processService,
     scenarioActivityRepository,
-    processResolverByProcessingType,
+    processResolverByProcessingType(),
     dbioRunner,
   )