+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/ProcessStateDefinitionServiceSpec.scala
@@ -5,7 +5,7 @@ import com.typesafe.config.ConfigFactory
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import pl.touk.nussknacker.engine.{JobsRecoverySettings, MetaDataInitializer}
-import pl.touk.nussknacker.engine.ProcessingTypeConfig.DeploymentManagerType
+import pl.touk.nussknacker.engine.ProcessingTypeConfig.{DeploymentManagerType, LimitsConfig}
 import pl.touk.nussknacker.engine.api.component.ComponentDefinition
 import pl.touk.nussknacker.engine.api.deployment._
 import pl.touk.nussknacker.engine.api.deployment.ProcessStateDefinitionManager.ScenarioStatusWithScenarioContext
@@ -15,6 +15,7 @@ import pl.touk.nussknacker.engine.api.process.{ProcessingType, Source, SourceFac
 import pl.touk.nussknacker.engine.deployment.EngineSetupName
 import pl.touk.nussknacker.engine.testing.LocalModelData
 import pl.touk.nussknacker.security.Permission
+import pl.touk.nussknacker.test.config.ConfigWithScalaVersion
 import pl.touk.nussknacker.test.mock.MockDeploymentManager
 import pl.touk.nussknacker.test.utils.domain.TestFactory.modelDependencies
 import pl.touk.nussknacker.test.utils.domain.TestProcessingTypeDataProviderFactory
@@ -187,6 +188,7 @@ class ProcessStateDefinitionServiceSpec extends AnyFunSuite with Matchers {
       category: String
   ): ProcessingTypeData = {
     val mockManager = MockDeploymentManager.create(
+      config = ConfigWithScalaVersion.StreamingProcessTypeConfig,
       customProcessStateDefinitionManager = Some(stateDefinitionManager)
     )
     val deploymentData = new DeploymentData(
@@ -207,7 +209,8 @@ class ProcessStateDefinitionServiceSpec extends AnyFunSuite with Matchers {
       ),
       deploymentData,
       category = category,
-      componentDefinitionExtractionMode = modelDependencies.componentDefinitionExtractionMode
+      limitsConfig = LimitsConfig.default,
+      componentDefinitionExtractionMode = modelDependencies.componentDefinitionExtractionMode,
     )
   }