+++ engine/flink/executor/src/main/scala/pl/touk/nussknacker/engine/process/compiler/FlinkProcessCompilerDataFactory.scala
@@ -1,13 +1,6 @@
 package pl.touk.nussknacker.engine.process.compiler
 
-import com.typesafe.config.Config
-import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment
-import pl.touk.nussknacker.engine.{
-  CustomProcessValidatorLoader,
-  ModelData,
-  RuntimeMode,
-  ScenarioCompilationDependencies
-}
+import pl.touk.nussknacker.engine.{CustomProcessValidatorLoader, ModelConfig, ModelData, RuntimeMode}
 import pl.touk.nussknacker.engine.ModelData.ExtractDefinitionFun
 import pl.touk.nussknacker.engine.api.{JobData, MetaData, ProcessListener, ProcessVersion}
 import pl.touk.nussknacker.engine.api.component.{
@@ -15,7 +8,6 @@ import pl.touk.nussknacker.engine.api.component.{
   DesignerWideComponentId,
   NodesDeploymentData
 }
-import pl.touk.nussknacker.engine.api.definition.EngineScenarioCompilationDependencies
 import pl.touk.nussknacker.engine.api.dict.EngineDictRegistry
 import pl.touk.nussknacker.engine.api.process.{ProcessConfigCreator, ProcessObjectDependencies}
 import pl.touk.nussknacker.engine.compile._
@@ -44,7 +36,7 @@ import scala.concurrent.duration.FiniteDuration
 class FlinkProcessCompilerDataFactory(
     creator: ProcessConfigCreator,
     extractModelDefinition: ExtractDefinitionFun,
-    modelConfig: Config,
+    modelConfig: ModelConfig,
     runtimeMode: RuntimeMode,
     configsFromProviderWithDictionaryEditor: Map[DesignerWideComponentId, ComponentAdditionalConfig],
     nodesData: NodesDeploymentData,
@@ -74,23 +66,24 @@ class FlinkProcessCompilerDataFactory(
       usedNodes: UsedNodes,
       userCodeClassLoader: ClassLoader
   ): FlinkProcessCompilerData = {
-    val modelDependencies = ProcessObjectDependencies.withConfig(modelConfig)
+    val modelDependencies = ProcessObjectDependencies.withConfig(modelConfig.underlyingConfig)
 
     // TODO: this should be somewhere else?
-    val timeout = modelConfig.as[FiniteDuration]("timeout")
+    val timeout = modelConfig.underlyingConfig.as[FiniteDuration]("timeout")
 
     // TODO: should this be the default?
     val asyncExecutionContextPreparer = creator
       .asyncExecutionContextPreparer(modelDependencies)
       .getOrElse(
-        modelConfig.as[DefaultServiceExecutionContextPreparer]("asyncExecutionConfig")
+        modelConfig.underlyingConfig.as[DefaultServiceExecutionContextPreparer]("asyncExecutionConfig")
       )
     val defaultListeners = prepareDefaultListeners(usedNodes) ++ creator.listeners(modelDependencies)
     val listenersToUse   = adjustListeners(defaultListeners, modelDependencies)
 
     val (definitionWithTypes, dictRegistry) = definitions(modelDependencies, userCodeClassLoader)
 
-    val customProcessValidator = CustomProcessValidatorLoader.loadProcessValidators(userCodeClassLoader, modelConfig)
+    val customProcessValidator =
+      CustomProcessValidatorLoader.loadProcessValidators(userCodeClassLoader, modelConfig.underlyingConfig)
     val compilerData =
       ProcessCompilerData.prepare(
         JobData(metaData, processVersion),