+++ extensions-api/src/main/scala/pl/touk/nussknacker/engine/ProcessingTypeConfig.scala
@@ -1,8 +1,11 @@
 package pl.touk.nussknacker.engine
 
-import com.typesafe.config.Config
-import pl.touk.nussknacker.engine.ProcessingTypeConfig.DeploymentManagerType
+import com.typesafe.config.{Config, ConfigException}
+import pl.touk.nussknacker.engine.ProcessingTypeConfig.{DeploymentManagerType, LimitsConfig}
+import pl.touk.nussknacker.engine.ProcessingTypeConfig.LimitsConfig.MaxActiveScenariosCount
 import pl.touk.nussknacker.engine.deployment.EngineSetupName
+
+import scala.util.{Failure, Success, Try}
 
 case class ProcessingTypeConfig(
     deploymentManagerType: DeploymentManagerType,
@@ -10,7 +13,8 @@ case class ProcessingTypeConfig(
     classPath: List[String],
     deploymentConfig: Config,
     modelConfig: ConfigWithUnresolvedVersion,
-    category: String
+    category: String,
+    limits: LimitsConfig
 )
 
 object ProcessingTypeConfig {
@@ -24,10 +28,31 @@ object ProcessingTypeConfig {
       config.resolved.as[List[String]]("modelConfig.classPath"),
       config.resolved.getConfig("deploymentConfig"),
       config.getConfig("modelConfig"),
-      config.resolved.as[String]("category")
+      config.resolved.as[String]("category"),
+      readLimitsConfig(config.resolved)
     )
   }
 
-  case class DeploymentManagerType(value: String)
+  private def readLimitsConfig(config: Config) = {
+    Try(config.getConfig("limits")) match {
+      case Success(limitsConfig) =>
+        LimitsConfig(
+          limitsConfig.getAs[Int]("maxActiveScenariosCount").map(MaxActiveScenariosCount.apply)
+        )
+      case Failure(_: ConfigException.Missing) =>
+        LimitsConfig.default
+      case Failure(ex) =>
+        throw ex
+    }
+  }
+
+  final case class DeploymentManagerType(value: String) extends AnyVal
+  final case class LimitsConfig(maxActiveScenariosCount: Option[MaxActiveScenariosCount])
+
+  object LimitsConfig {
+    final case class MaxActiveScenariosCount(value: Int) extends AnyVal
+
+    val default = LimitsConfig(maxActiveScenariosCount = None)
+  }
 
 }